Separate pages; move middleware to server.js
