+++ apps/prairielearn/src/pages/navbarCourseInstanceSwitcher/navbarCourseInstanceSwitcher.html.ts
@@ -2,34 +2,10 @@ import { html } from '@prairielearn/html';
 
 import { type CourseInstance } from '../../lib/db-types';
 import { idsEqual } from '../../lib/id';
-import { Course } from '../../lib/db-types';
 
 // TODO: Once our `course_instances_with_staff_access` sproc is returning full
 // course instance rows, use the full `CourseInstance` type here.
 type CourseInstanceWithShortName = Pick<CourseInstance, 'id' | 'short_name'>;
-
-export function NavbarCourseSwitcher({
-  courses,
-  current_course_id,
-  plainUrlPrefix,
-}: {
-  courses: Course[];
-  current_course_id: string;
-  plainUrlPrefix: string;
-}) {
-  return html`
-    ${courses.map((c) => {
-      return html`
-        <a
-          class="dropdown-item ${idsEqual(c.id, current_course_id) ? 'active' : ''}"
-          href="${plainUrlPrefix}/course/${c.id}/course_admin"
-        >
-          ${c.short_name}
-        </a>
-      `;
-    })}
-  `.toString();
-}
 
 export function NavbarCourseInstanceSwitcher({
   course_instances,
